+++ flydra2-mainbrain/src/lib.rs
@@ -294,7 +294,19 @@ pub async fn pre_run(
         let cal_file = anyhow::Context::with_context(std::fs::File::open(&cal_fname), || {
             format!("loading calibration {}", cal_fname.display())
         })?;
-        Some(flydra_mvg::FlydraMultiCameraSystem::<MyFloat>::from_flydra_xml(cal_file)?)
+
+        if cal_fname.extension() == Some(std::ffi::OsStr::new("json"))
+            || cal_fname.extension() == Some(std::ffi::OsStr::new("pymvg"))
+        {
+            // Assume any .json or .pymvg file is a pymvg file.
+            let system = mvg::MultiCameraSystem::<MyFloat>::from_pymvg_file_json(cal_file)?;
+            Some(flydra_mvg::FlydraMultiCameraSystem::<MyFloat>::from_system(
+                system, None,
+            ))
+        } else {
+            // Otherwise, assume it is a flydra xml file.
+            Some(flydra_mvg::FlydraMultiCameraSystem::<MyFloat>::from_flydra_xml(cal_file)?)
+        }
     } else {
         None
     };